+++ Svc/PolyDb/PolyDb.cpp
@@ -5,26 +5,26 @@
  *      Author: Timothy Canham
  */
 
-#include <Svc/PolyDb/PolyDbImpl.hpp>
+#include <Svc/PolyDb/PolyDb.hpp>
 #include <Fw/Types/Assert.hpp>
 #include <FpConfig.hpp>
 
 namespace Svc {
-    PolyDbImpl::PolyDbImpl(const char* name) : PolyDbComponentBase(name) {
+    PolyDb::PolyDb(const char* name) : PolyDbComponentBase(name) {
         // initialize all entries to stale
         for (NATIVE_INT_TYPE entry = 0; entry < Svc::PolyDbCfg::PolyDbEntry::NUM_CONSTANTS; entry++) {
             this->m_db[entry].status = MeasurementStatus::STALE;
         }
     }
 
-    void PolyDbImpl::init(NATIVE_INT_TYPE instance) {
+    void PolyDb::init(NATIVE_INT_TYPE instance) {
         PolyDbComponentBase::init(instance);
     }
 
     // If ports are no longer guarded, these accesses need to be protected from each other
     // If there are a lot of accesses, perhaps an interrupt lock could be used instead of guarded ports
 
-    void PolyDbImpl ::
+    void PolyDb ::
         getValue_handler(
             NATIVE_INT_TYPE portNum,
             const Svc::PolyDbCfg::PolyDbEntry& entry,
@@ -38,7 +38,7 @@ namespace Svc {
         val = this->m_db[entry.e].val;
     }
 
-    void PolyDbImpl ::
+    void PolyDb ::
         setValue_handler(
             NATIVE_INT_TYPE portNum,
             const Svc::PolyDbCfg::PolyDbEntry& entry,
@@ -52,7 +52,7 @@ namespace Svc {
         this->m_db[entry.e].val = val;
     }
 
-    PolyDbImpl::~PolyDbImpl() {
+    PolyDb::~PolyDb() {
     }
 
 }
